update (#1)
* Make sure both gpg-agent and dirmngr have picked up new configuration (#1387)

@hamidzr wrote #1335, which fixes #1333, a case of a running `gpg-agent` not picking up changes from GPG configuration. When testing, I noticed the configuration issue affects `dirmngr` as well. If there's a running `dirmngr`, provisioning fails when trying to update the keyring. This patch addresses both components.

`dirmngr` doesn't seem respond to HUP or `--reload`, so it needs to be killed and restarted.

Resolves #1333.

* - mobileconfig re-generated from apple configurator 2 (#1398)


    - .mobileconfig re-generated from Apple Configurator 2 (DeviceUniqueIdentifier key/value from VedorConfig and ProviderType packet-tunnel setting now removed)
    - uppercase UUIDs

* Remove "ec2:DescribeAlarms" from suggested IAM policy (#1407)

"ec2:DescribeAlarms" doesn't exist.

* Updated Documentation for Ubuntu 16.04, 17.10, and 18.04 (#1408)

There have been some issues with versions of Ubuntu. See Issue #1296 for information about problems with 16.04. 16.10 and 17.04 are now end of life (see https://wiki.ubuntu.com/Releases). I have updated the documentation (minor changes) to reflect 16.04, 17.10, and 18.04, which are currently supported.

* Ensure VPN profile names fit in Linux network device names (#1400)

There's a 15-character limit on the length of Linux network device names, and it would be nice to use the VPN profile names as interface names. A few of the wordlist words are 8 letters long. We can use two words as long as they're each 7 characters or shorter (leaving room for the separator).

* When checking syntax, don't search inside the venv directory (#1395)

Now that we're telling people to use ./venv as a place to put virtualenvs, don't search there when we're checking syntax.

* Don't print long apt keys to the log (#1412)

This is mostly for the 200-line Tor key, but the other (25-line) keys were hidden for consistency.

* Add WireGuard instructions for Android using the new QR code support. (#1413)

* WireGuard instructions update (#1417)

* Add WireGuard instructions for Android using the new QR code support.

* update wireguard fr instructions

* Update readme-fr (#1419)

* Update OpenVPN instuctions (#1420)

- update fr instructions to reflect changes made in #1408
- correct instructions to mention LZ4 compression instead of LZO

* Fix stunnel windows mirror download location (#1425)